backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
librosa==0.10.2
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.1
noisereduce==3.0.2
python-multipart==0.0.9
soundfile==0.12.1
pydantic==2.9.1
python-dotenv==1.0.1
supabase==2.7.2
openai==1.52.0
joblib==1.5.3
torch==2.5.1
